Add files via upload
I am writing to provide a description of the Amazon HTML, CSS Clone Project that I am currently working on. The project aims to recreate the Amazon website using HTML and CSS. This is a front-end development project that focuses on replicating the user interface of Amazon's website as closely as possible. The goal is to enhance my skills in HTML and CSS, while also gaining a deeper understanding of website design and development. The project will involve creating various pages of the website, including the homepage, product pages, and shopping cart. Each page will be built using HTML and styled using CSS. The project will also incorporate responsive design principles to ensure the website is accessible and visually appealing on a variety of devices. I am excited about this project as it allows me to apply and expand my knowledge in HTML and CSS. I believe it will be a valuable addition to my portfolio. Please feel free to reach out if you have any questions or would like further information. Best regards, Manasa Vybhavi.